pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "For the ETLV"
readme = "README.md"
authors = [
    { name = "Lameur", email = "96985489+Lameur@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "python-pptx>=1.0.2",
    "seaborn>=0.13.2",
]

[project.scripts]
app = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
